rre-dataset-generator/pyproject.toml
[project]
name = "Dataset-Generator"
version = "0.1.0"
description = "Dataset Generator project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "mypy-extensions>=1.1.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.10",
    "typing-extensions>=4.14.1",
    "types-PyYAML>=6.0.2",
    "types-requests>=2.32.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
files = ["src", "dataset_generator.py"]
exclude = "^tests/integration/"
# Strict but reasonable
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
show_error_codes = true
pretty = true

# Silence libs without stubs
[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_core.*",
    "langchain_openai.*",
    "langchain_google_genai.*",
    "dotenv.*"
]
ignore_missing_imports = true

# Allow untyped functions in tests for speed
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

[tool.pydantic-mypy]
init_typed = true
warn_required_dynamic_aliases = true
